Conflicts:
	EngineBase/Collections/IIndex.cs

Former-commit-id: eb26f40246ee644d2756f568cc9f03365ad200ec

+++ SpaceBase/Graphics/Planet.cs
@@ -130,7 +130,7 @@ namespace Space.Graphics
             _effect.Parameters["EmbossScale"].SetValue(1f / _width);
 
             _effect.Parameters["TextureOffset"].SetValue(_surfaceRotation * ((float)_gameTime.TotalGameTime.TotalSeconds / _width));
-            _effect.Parameters["TextureScale"].SetValue((float)_surface.Width / (2f * _width));
+            _effect.Parameters["TextureScale"].SetValue(_surface.Width / (2f * _width));
         }
 
         #endregion
